refactor: separate admin resource management APIs
- Move theme and time create/delete APIs under /api/v1/admin
- Remove userName-based admin validation from request bodies
- Remove DELETE request bodies and unauthorized exception handling
- Update controller tests to use admin endpoints for setup
