+++ src/main/java/roomescape/controller/AdminReservationTimeController.java
@@ -1,0 +1,39 @@
+package roomescape.controller;
+
+import java.net.URI;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.ReservationTime;
+import roomescape.dto.request.ReservationTimeCreateRequest;
+import roomescape.service.ReservationTimeService;
+
+@RequestMapping("/api/v1/admin/times")
+@RestController
+public class AdminReservationTimeController {
+
+    private final ReservationTimeService reservationTimeService;
+
+    public AdminReservationTimeController(ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
+    }
+
+    @PostMapping
+    public ResponseEntity<ReservationTime> createReservationTime(
+            @RequestBody ReservationTimeCreateRequest reservationTimeCreateRequest) {
+        ReservationTime savedReservationTime = reservationTimeService.createReservationTime(
+                reservationTimeCreateRequest.startAt());
+        return ResponseEntity.created(URI.create("/api/v1/admin/times/" + savedReservationTime.getId()))
+                .body(savedReservationTime);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteReservationTime(@PathVariable Long id) {
+        reservationTimeService.deleteReservationTime(id);
+        return ResponseEntity.noContent().build();
+    }
+}
