+++ src/main/java/roomescape/controller/AdminThemeController.java
@@ -1,0 +1,41 @@
+package roomescape.controller;
+
+import java.net.URI;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.domain.Theme;
+import roomescape.dto.request.ThemeCreateRequest;
+import roomescape.service.ThemeService;
+
+@RestController
+@RequestMapping("/api/v1/admin/themes")
+public class AdminThemeController {
+
+    private final ThemeService themeService;
+
+    public AdminThemeController(ThemeService themeService) {
+        this.themeService = themeService;
+    }
+
+    @PostMapping
+    public ResponseEntity<Theme> createTheme(
+            @RequestBody ThemeCreateRequest themeCreateRequest) {
+        Theme theme = themeService.createTheme(
+                themeCreateRequest.name(),
+                themeCreateRequest.description(),
+                themeCreateRequest.imgUrl()
+        );
+        return ResponseEntity.created(URI.create("/api/v1/admin/themes/" + theme.getId())).body(theme);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteTheme(@PathVariable Long id) {
+        themeService.deleteTheme(id);
+        return ResponseEntity.noContent().build();
+    }
+}
