--- src/main/java/roomescape/dto/request/ReservationTimeDeleteRequest.java
@@ -1,6 +1,0 @@
-package roomescape.dto.request;
-
-public record ReservationTimeDeleteRequest(
-        String userName
-) {
-}
