+++ src/main/java/roomescape/dto/request/ThemeCreateRequest.java
@@ -3,7 +3,6 @@ package roomescape.dto.request;
 public record ThemeCreateRequest(
         String name,
         String description,
-        String imgUrl,
-        String userName
+        String imgUrl
 ) {
 }
