--- src/main/java/roomescape/dto/request/ThemeDeleteRequest.java
@@ -1,6 +1,0 @@
-package roomescape.dto.request;
-
-public record ThemeDeleteRequest(
-        String userName
-) {
-}
