--- src/main/java/roomescape/exception/UnauthorizedException.java
@@ -1,4 +1,0 @@
-package roomescape.exception;
-
-public class UnauthorizedException extends RuntimeException {
-}
