+++ src/test/java/roomescape/service/ReservationTimeServiceTest.java
@@ -4,14 +4,12 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-import java.time.LocalTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.dao.DataIntegrityViolationException;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.exception.ReservationTimeInUseException;
 import roomescape.exception.ReservationTimeNotFoundException;
-import roomescape.exception.UnauthorizedException;
 
 public class ReservationTimeServiceTest {
 
@@ -25,25 +23,11 @@ public class ReservationTimeServiceTest {
     }
 
     @Test
-    void 관리자가_아니면_시간을_생성할_수_없다() {
-        assertThatThrownBy(() -> reservationTimeService.createReservationTime(
-                LocalTime.parse("10:00"), "정콩이")).
-                isInstanceOf(UnauthorizedException.class);
-    }
-
-    @Test
-    void 관리자가_아니면_시간을_삭제할_수_없다() {
-        when(reservationTimeDao.delete(1L)).thenReturn(1);
-        assertThatThrownBy(() -> reservationTimeService.deleteReservationTime(1L, "정콩이"))
-                .isInstanceOf(UnauthorizedException.class);
-    }
-
-    @Test
     void 예약이_있는_시간을_삭제할_수_없다() {
         when(reservationTimeDao.delete(1L))
                 .thenThrow(new DataIntegrityViolationException("foreign key violation"));
 
-        assertThatThrownBy(() -> reservationTimeService.deleteReservationTime(1L, "ADMIN"))
+        assertThatThrownBy(() -> reservationTimeService.deleteReservationTime(1L))
                 .isInstanceOf(ReservationTimeInUseException.class);
     }
 
@@ -51,7 +35,7 @@ public class ReservationTimeServiceTest {
     void 존재하지않는_시간을_삭제할_수_없다() {
         when(reservationTimeDao.delete(1L)).thenReturn(0);
 
-        assertThatThrownBy(() -> reservationTimeService.deleteReservationTime(1L, "ADMIN"))
+        assertThatThrownBy(() -> reservationTimeService.deleteReservationTime(1L))
                 .isInstanceOf(ReservationTimeNotFoundException.class);
     }
 }
