+++ src/test/java/roomescape/service/ThemeServiceTest.java
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import roomescape.dao.ThemeDao;
 import roomescape.exception.ThemeNotFoundException;
-import roomescape.exception.UnauthorizedException;
 
 public class ThemeServiceTest {
 
@@ -20,21 +19,8 @@ public class ThemeServiceTest {
     }
 
     @Test
-    void 관리자가_아니면_테마를_생성할_수_없다() {
-        assertThatThrownBy(() -> themeService.createTheme(
-                "공포", "설명", "image.jpg", "USER"))
-                .isInstanceOf(UnauthorizedException.class);
-    }
-
-    @Test
-    void 관리자가_아니면_테마를_삭제할_수_없다() {
-        assertThatThrownBy(() -> themeService.deleteTheme(1L, "USER"))
-                .isInstanceOf(UnauthorizedException.class);
-    }
-
-    @Test
     void 없는_테마는_삭제할_수_없다() {
-        assertThatThrownBy(() -> themeService.deleteTheme(1L, "ADMIN"))
+        assertThatThrownBy(() -> themeService.deleteTheme(1L))
                 .isInstanceOf(ThemeNotFoundException.class);
     }
 }
